feat(socket): add error handler to events
