+++ src/errors/socket-error-handler.ts
@@ -1,15 +1,3 @@
-import { ExtendedError } from 'socket.io/dist/namespace';
-
-export const socketErrorHandler = (
-  err: unknown,
-  next: (err?: ExtendedError | undefined) => void
-) => {
-  let message = "Error occurred in Socket's middleware: ";
-  if (err && typeof err === 'object' && 'message' in err) {
-    message += err.message;
-  } else {
-    message += '<no details>';
-  }
-  const extendedErr: ExtendedError = { name: 'SocketMiddlewareError', message };
-  next(extendedErr);
+export const socketErrorHandler = (err: unknown) => {
+  console.log("Error occurred in Socket's event:\n", JSON.stringify(err, null, 2));
 };
